reorganize scripts into phase4bis
