+++ phase4/extra_visuals.py
@@ -29,7 +29,7 @@ import seaborn as sns
 import yaml
 from sklearn.cluster import KMeans
 
-from phase4_functions import (
+from phase4.functions import (
     load_datasets,
     prepare_data,
     select_variables,
