+++ phase4/generate_csv.py
@@ -21,7 +21,7 @@ import yaml
 from scipy.spatial.distance import pdist, squareform
 
 import phase4
-from phase4_functions import (
+from phase4.functions import (
     load_datasets,
     prepare_data,
     select_variables,
