+++ phase4bis/export_famd_contributions.py
