+++ phase4bis/export_famd_inertias.py
