+++ phase4bis/export_mfa_group_contributions.py
