+++ phase4bis/export_mfa_inertias.py
