+++ phase4bis/export_pca_inertias.py
