+++ phase4bis/famd_full_analysis.py
